[PATCH 2/9] libao: Add a detect() function
Most Linux distros use pulseaudio by now, and libao is the best way to use
pulseaudio from timidity. This patch adds a detect function to libao, which
will make libao autodetect succeed only when pulse is available, so that
we can prefer libao in the probe order to get pulse, without also getting
libao on systems where we should be using alsa directly.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
---
 timidity/ao_a.c |   31 ++++++++++++++++++++++++++++++-
 1 files changed, 30 insertions(+), 1 deletions(-)