Revert old spectogram workaround
An old change of mine to miditrace.c was made to allow the spectrogram
(usually attached to 'g' key) to work in trace mode (can't remember if
the problem was in the xaw interface alone or not). Apparently, this
caused problems on the mac interface which led to the ifdef next door
(didn't notice this at the time). Now, local testing shows it's no
longer necessary on unix. So I suggest reverting it for simpler code.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>